ampliacion de la informacion retornada por me, completado del recurso consejo_role_user y un poco de limpieza del codigo
En la rama rolesIdvm
Cambios a ser confirmados:
	modificado:     app/Http/Controllers/AuthController.php

    se expandio la informacion retornada por me

    se expandio las funcionalidades del recurso consejo_role_user
    y se reforzo la seguridad agregaron mas validaciones
	modificado:     app/Http/Controllers/ConsejoRoleUserController.php

    se reforzo la seguridad agregando algunas validaciones al metodo de retornar
    modificado:     app/Http/Controllers/UserController.php

    se agrego un helper para facilitar la obtencion de un rol desde un determinado usuario
	modificado:     app/helpers.php

    se quitaron los midlewares que no se estaban
    usando en la parte de roles y permisos y users
    modificado:     routes/api.php

    se limpio el codigo
	modificado:     app/Http/Controllers/ConsejoPuntoController.php
	borrado:        app/Http/Controllers/ConsejoUsuarioController.php
	modificado:     app/Models/Agenda.php
	modificado:     app/Models/Consejo.php
	renombrado:     app/Models/Consejo_Punto.php -> app/Models/ConsejoPunto.php
	modificado:     database/seeders/PermisoRoleSeeder.php

+++ app/Http/Controllers/UserController.php
@@ -25,8 +25,34 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $user = User::create($request->all());
-        return $user;
+        if(!userCan(auth()->user()['id'],"agregar usuario")){
+            return response()->json(
+                [
+                    'message' => 'No tienes el permiso para hacer esto',
+                ]
+                ,401);
+        }
+
+        $validator = Validator::make(request()->all(), [
+            'name' => 'required|string|max:50',
+            'cedula' => 'required|numeric|digits_between:6,8|unique:users,cedula',
+            'email' => 'required|string|email|max:255|unique:users',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors()->toJson(), 400);
+        }
+
+        $user = User::create(array_merge (
+            $validator->validate(),
+            ['password' => bcrypt($request->cedula)]
+        ));
+
+        return response()->json([
+            'message' => 'User created successfully',
+            'data' => $user
+        ]);
+    
     }
 
     /**
@@ -49,6 +75,14 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(auth()->user()['id']!=$id && !userCan(auth()->user()['id'],"actualizar usuario")){
+            return response()->json(
+                [
+                    'message' => 'No tienes el permiso para hacer esto',
+                ]
+                ,401);
+        }
+
         $user = User::find($id);
         $user->update($request->all());
         return $user;
@@ -62,8 +96,18 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+
+        if(auth()->user()['id']!=$id && !userCan(auth()->user()['id'],"eliminar usuario")){
+            return response()->json(
+                [
+                    'message' => 'No tienes el permiso para hacer esto',
+                ]
+                ,401);
+        }
+
         $user = User::find($id);
         $user->delete();
         return $user;
     }
+
 }
